# docs/openapi/innovation-sandbox-api.yaml
openapi: 3.0.0
info:
  title: Innovation Sandbox on AWS
  description: API documentation for the Innovation Sandbox on AWS solution.
  version: 1.3.0
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "{apiBaseUrl}"
    description: Base URL used to make API requests
    variables:
      apiBaseUrl:
        default: https://example.com/api
        description: Base URL used to make API requests
...
security:
  - sigV4: []
    isbIdentity: [] # human callers (Identity Pool roles)
  - sigV4: [] # M2M callers (per-client assumed roles)
components:
  securitySchemes:
    sigV4:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: awsSigv4
      description: >-
        AWS Signature Version 4 over the `execute-api` service. Required on
        every request: the API uses IAM authorization
        (`AuthorizationType.IAM`), and there is no bearer-token or Cognito
        authorizer. The caller type is determined from the IAM principal.
    isbIdentity:
      type: apiKey
      name: x-isb-identity
      in: header
      description: >-
        Cognito ID token, verified via JWKS. Carries the claims used for RBAC.
        Required for human callers (Identity Pool roles). M2M callers (per-client
        assumed IAM roles) derive their role from the assumed-role ARN and must
        NOT send this header; a request presenting both an M2M principal and
        `x-isb-identity` is rejected with 400.
  schemas:
...
    SuccessResponseBody:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
        - type: object
          required:
            - status
            - data
          properties:
            status:
              type: string
              enum:
                - success
            data:
              type: object
...
    PaginatedResults:
      type: object
      required:
        - result
      properties:
        result:
          type: array
        nextPageIdentifier:
          type: string
          example: "eyJsYXN0RXZhbHVhdGVkS2V5Ijp7InV1aWQiOnsicyI6IjEyMzQ1Njc4LTkwYWItY2RlZi0xMjM0LTU2Nzg5MGFiY2RlZiJ9fX0="
    AwsAccountId:
      type: string
      description: AWS account ID
      example: "123456789012"
...
    Account:
      type: object
      required:
        - awsAccountId
        - status
      properties:
        awsAccountId:
          $ref: "#/components/schemas/AwsAccountId"
        email:
          type: string
          format: email
          description: Email address associated with the AWS account
        name:
          type: string
          description: Friendly name of the account
          maxLength: 50
        status:
          $ref: "#/components/schemas/AccountStatus"
        driftAtLastScan:
          $ref: "#/components/schemas/DriftAtLastScan"
        cleanupExecutionContext:
          type: object
          description: Context for an in-progress cleanup execution
          properties:
            stateMachineExecutionArn:
              type: string
              description: ARN of the Step Functions execution performing cleanup
            stateMachineExecutionStartTime:
              type: string
              format: date-time
              description: When the cleanup execution started
        meta:
          $ref: "#/components/schemas/MetaData"
...
  parameters:
    PageIdentifier:
      name: pageIdentifier
      in: query
      required: false
      schema:
        type: string
      description: Base64 encoded token for pagination
    MaxResults:
      name: maxResults
      in: query
      required: false
      schema:
        type: integer
      description: >-
        Maximum number of items to return per page. The legacy `pageSize`
        parameter is unsupported: most endpoints ignore it and apply their
        default, while `/leases/shared` rejects it as an unknown parameter.
...
  responses:
    BadRequestResponse:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FailResponseBody"
          example:
            status: fail
            data:
              errors: [{ "message": "example error message" }]
...
paths:
...
  /accounts:
    get:
      tags:
        - accounts
      summary: Get paginated list of all accounts
      parameters:
        - $ref: "#/components/parameters/PageIdentifier"
        - $ref: "#/components/parameters/MaxResults"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PaginatedResults"
                          - type: object
                            properties:
                              result:
                                type: array
                                items:
                                  $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags:
        - accounts
      summary: Register account with account pool
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - awsAccountId
              properties:
                awsAccountId:
                  $ref: "#/components/schemas/AwsAccountId"
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /accounts/{awsAccountId}:
    parameters:
      - $ref: "#/components/parameters/AwsAccountId"
    get:
      tags:
        - accounts
      summary: Get account by account id
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/ResourceNotFoundResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /accounts/{awsAccountId}/retryCleanup:
    parameters:
      - $ref: "#/components/parameters/AwsAccountId"
    post:
      tags:
        - accounts
      summary: Retry cleanup for an account
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/ResourceNotFoundResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /accounts/{awsAccountId}/eject:
    parameters:
      - $ref: "#/components/parameters/AwsAccountId"
    post:
      tags:
        - accounts
      summary: Eject an account from the sandbox
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/ResourceNotFoundResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /accounts/{awsAccountId}/quarantine:
    parameters:
      - $ref: "#/components/parameters/AwsAccountId"
    post:
      tags:
        - accounts
      summary: Manually quarantine an account
      description: >
        Terminates any active lease, revokes leaseholder SSO access, and moves the account
        to the Quarantine OU. Eligibility: Available, Active, or Frozen accounts.
        Returns 409 if the account is already quarantined or has an active cleanup in progress.
        Admin role only.
      responses:
        "200":
          description: Account successfully quarantined
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/ResourceNotFoundResponse"
        "409":
          description: Account is already quarantined or cleanup is in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponseBody"
              examples:
                alreadyQuarantined:
                  value:
                    status: fail
                    data:
                      errors:
                        - message: "Account is already quarantined."
                cleanupInProgress:
                  value:
                    status: fail
                    data:
                      errors:
                        - message: "Account cannot be quarantined while cleanup is in progress."
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /accounts/{awsAccountId}/cleanup-reports:
    parameters:
      - $ref: "#/components/parameters/AwsAccountId"
    get:
      tags:
        - accounts
      summary: List cleanup reports for an account
      description: Returns recent cleanup reports for the specified account, newest first. Defaults to the last 5 reports. Admin role only.
      parameters:
        - $ref: "#/components/parameters/PageIdentifier"
        - name: maxResults
          in: query
          required: false
          schema:
            type: integer
            default: 5
          description: >-
            Number of reports to return (default 5). The legacy `pageSize`
            parameter is ignored and the default is applied.
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PaginatedResults"
                          - type: object
                            properties:
                              result:
                                type: array
                                items:
                                  $ref: "#/components/schemas/CleanupReport"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/ResourceNotFoundResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /accounts/{awsAccountId}/skipCooldown:
    parameters:
      - $ref: "#/components/parameters/AwsAccountId"
    post:
      tags:
        - accounts
      summary: Skip the post-cleanup cooldown for an account
      description: >
        Immediately resumes the cleanup durable execution, skipping the remaining
        post-cleanup cooldown period and returning the account to the Available pool.
        The skipCooldownCallbackId is resolved server-side from the latest cleanup
        report and is never exposed to the client. Admin role only.
      responses:
        "200":
          description: Cooldown skipped successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "Cooldown skipped successfully."
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/ResourceNotFoundResponse"
        "409":
          description: Account is not currently in cooldown, or cooldown has already completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponseBody"
              example:
                status: fail
                data:
                  errors:
                    - message: "Account is not in an active cooldown."
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /accounts/unregistered:
    get:
      tags:
        - accounts
      summary: Get list of unregistered accounts in Entry OU
      parameters:
        - $ref: "#/components/parameters/PageIdentifier"
        - $ref: "#/components/parameters/MaxResults"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PaginatedResults"
                          - type: object
                            properties:
                              result:
                                type: array
                                items:
                                  $ref: "#/components/schemas/UnregisteredAccount"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"

  # Configuration endpoints
...